pyproject.toml
[tool.poetry]
name = "sushi"
version = "0.1.0"
description = ""
authors = ["Rustam <rustamelutingol@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8"
django-debug-toolbar = "^2.0"
django-phonenumber-field = "^3.0"
django-tz-detect = "^0.2.10"
django-registration = "^3.0"
phonenumbers = "^8.10"
django-widget-tweaks = "^1.4"
wagtail = "^2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
